pyproject.toml
[project]
name = "xyte-mcp"
version = "1.1.0"
description = "MCP server for Xyte organization API"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
readme = "README.md"

# MCP code and Dockerfile both target 3.11
requires-python = ">=3.11"

# ───────── FULL PACKAGE LIST (runtime + tooling) ─────────
dependencies = [
    # ── MCP core & HTTP stack ──
    "mcp[cli]>=1.1.0",
    "starlette>=0.36",
    "uvicorn[standard]>=0.29",

    # ── Auth helpers ──
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7",

    # ── HTTP client & caching ──
    "httpx>=0.27.0",
    "cachetools>=5.3.0",

    # ── Redis / Streams / Lua ──
    "redis[async]>=5.0",

    # ── Postgres async ORM ──
    "sqlmodel>=0.0.12",
    "asyncpg>=0.29",

    # ── Background tasks ──
    "celery[redis]>=5.3",

    # ── YAML & JWT utilities ──
    "pyyaml>=6.0",
    "pyjwt>=2.8",

    # ── Observability ──
    "prometheus-client>=0.19.0",
    "opentelemetry-sdk>=1.25.0",

    # ── Config / env ──
    "python-dotenv>=1.0.0",

    # ── Data validation ──
    "pydantic>=2.0.0",

    # ── Tooling & QA (needed because install runs only once) ──
    "pytest>=8.2",
    "ruff>=0.4",
    "pyright>=1.1",
    "safety>=3.2",
    "mkdocs-material>=9.5",
]

[project.scripts]
serve = "xyte_mcp:serve"
xyte-mcp = "xyte_mcp.server:main"

[project.urls]
Homepage = "https://github.com/xyte/xyte-mcp"

# ── Optional-dependencies table can stay, but Codex won’t honor it ──
[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "ruff>=0.4",
    "pyright>=1.1",
    "safety>=3.2",
    "mkdocs-material>=9.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xyte_mcp"]

[tool.ruff]
line-length = 100
